refactor: image name and location
