+++ src/app/_components/MahasiswaSidebar.tsx
@@ -14,7 +14,7 @@ import { MagnifyingGlassIcon } from '@radix-ui/react-icons';
 import { signOut } from "next-auth/react";
 import {type Session} from "next-auth";
 // Assets Import
-import LogoAnmategra from "/public/images/logo-anmategra.png"
+import LogoAnmategra from "/public/images/logo/anmategra-logo-full.png"
 
 const MahasiswaSidebar = ({ session }: { session: Session | null }) => {
   const [searchQuery, setSearchQuery] = useState('');
