+++ src/app/authentication/page.tsx
@@ -7,9 +7,9 @@ import Image from 'next/image'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '~/components/ui/card'
 import { Button } from '~/components/ui/button'
 // Asset Import
-import Logo from '../../../public/images/logo-anmategra.png'
-import Google from '../../../public/images/google.webp'
-import Microsoft from '../../../public/images/microsoft.webp'
+import Logo from '../../../public/images/logo/anmategra-logo-full.png'
+import Google from 'public/icons/google-icon.webp'
+import Microsoft from 'public/icons/microsoft-icon.webp'
 
 import { signIn } from "next-auth/react";
