+++ src/app/lembaga/profil-lembaga/[lembagaId]/page.tsx
@@ -3,9 +3,9 @@ import React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
 // Asset Import
-import DummyFotoLembaga from '../../../../../public/images/logo-hmif.png'
-import DummyFotoEvent from 'public/placeholder/kegiatan thumbnail.png'
-import LogoHMIFKecil from 'public/placeholder/logo if.png'
+import DummyFotoLembaga from 'public/images/logo/hmif-logo.png'
+import DummyFotoEvent from 'public/placeholder/kegiatan-thumbnail.png'
+import LogoHMIFKecil from 'public/placeholder/logo-hmif.png'
 // Components Import
 import { Card } from '~/components/ui/card'
 import { Badge } from '~/components/ui/badge'
